requirements.txt
flask==3.0.0
flask-socketio==5.3.5
opencv-contrib-python
google-generativeai==0.3.2
python-dotenv==1.0.0
numpy
pillow
eventlet==0.35.2

